Create Thermostat Setpoints: V-Momentary Dimmers v1
This Thermostat SetPoint app uses a custom Virtual Momentary Dimmer Switches.  These new switches are used in place of the traditional dimmer, since the on/off state is not as critical for this app, since the focus is on the level slider.  It should also help keep button state in sync, when check the setpoint state of the thermostat.