Update and rename Thermostat Setpoints: V-Momentary Dimmers v2 to Thermostat Setpoints: V-Momentary Dimmers v2.1
Modified App: 
Added option to chose between the normal setpoint method and the quickSet Method.  This quickSet method is
used in some thermostat device types.  In those device types there is a delay build into trigger the thermostat change via
the switch.  The delay is longer on normal setting and greatly reduced on the quickSet Method.